Fix configurations
Fixed plugin configurations using new command service
