+++ src/TobysBot.Misc/MiscPlugin.cs
@@ -1,0 +1,8 @@
+﻿namespace TobysBot.Misc;
+
+public class MiscPlugin : IPluginRegistration
+{
+    public string Id => "misc";
+    public string Name => "Miscellanious";
+    public string Description => "Assorted random commands.";
+}
