--- src/TobysBot.Music/Commands/MusicPlugin.cs
@@ -1,12 +1,0 @@
-﻿using Discord.Commands;
-using TobysBot.Commands;
-
-namespace TobysBot.Music.Commands;
-
-[Name("Music")]
-[Id("music")]
-[Summary("Listen to music on Discord with Toby's Bot")]
-public partial class MusicPlugin : PluginBase
-{
-    
-}
