+++ src/TobysBot.Music/MusicPlugin.cs
@@ -1,0 +1,8 @@
+﻿namespace TobysBot.Music;
+
+public class MusicPlugin : IPluginRegistration
+{
+    public string Id => "music";
+    public string Name => "Music";
+    public string Description => "Listen to music on Discord with Toby's Bot";
+}
