--- src/TobysBot.Voice/Commands/VoicePlugin.cs
@@ -1,12 +1,0 @@
-﻿using Discord.Commands;
-using TobysBot.Commands;
-
-namespace TobysBot.Voice.Commands;
-
-[Name("Voice")]
-[Id("voice")]
-[Summary("Manage Toby's Bot in the voice channel")]
-public partial class VoicePlugin : PluginBase
-{
-    
-}
