+++ src/TobysBot.Voice/VoicePlugin.cs
@@ -1,0 +1,8 @@
+﻿namespace TobysBot.Voice;
+
+public class VoicePlugin : IPluginRegistration
+{
+    public string Id => "voice";
+    public string Name => "Voice";
+    public string Description => "Manage Toby's Bot in the voice channel";
+}
