+++ src/TobysBot/TobysBotHostedService.cs
@@ -4,6 +4,7 @@ using Discord.WebSocket;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using TobysBot.Commands;
 using TobysBot.Configuration;
 using TobysBot.Events;
 using TobysBot.Extensions;
@@ -13,13 +14,13 @@ namespace TobysBot;
 public class TobysBotHostedService : IHostedService
 {
     private readonly DiscordSocketClient _client;
-    private readonly CommandService _commands;
+    private readonly ICommandService _commands;
     private readonly IEventService _events;
     private readonly IServiceProvider _services;
     private readonly ILogger<TobysBotHostedService> _logger;
     private readonly TobysBotOptions _options;
 
-    public TobysBotHostedService(DiscordSocketClient client, CommandService commands, IEventService events, IServiceProvider services, IOptions<TobysBotOptions> options, ILogger<TobysBotHostedService> logger)
+    public TobysBotHostedService(DiscordSocketClient client, ICommandService commands, IEventService events, IServiceProvider services, IOptions<TobysBotOptions> options, ILogger<TobysBotHostedService> logger)
     {
         _client = client;
         _commands = commands;
@@ -52,8 +53,7 @@ public class TobysBotHostedService : IHostedService
 
         try
         {
-            var modules = await _commands.InstallCommandsAsync(_services);
-            await _client.InstallSlashCommandsAsync(modules);
+            await _commands.InstallCommandsAsync();
         }
         catch (Exception ex)
         {
